+++ test/unit/mysql_db_admin/data_out.py
@@ -22,15 +22,14 @@ import mock
 
 # Local
 sys.path.append(os.getcwd())
-import lib.gen_libs as gen_libs
-import mysql_db_admin
-
-import version
+import mysql_db_admin                           # pylint:disable=E0401,C0413
+import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class MailTest(object):
+class MailTest():
 
     """Class:  MailTest
 
@@ -111,8 +110,6 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_mongo_error
-        test_mongo
         test_outfile_mode2
         test_outfile_mode
         test_outfile
@@ -164,42 +161,8 @@ class UnitTest(unittest.TestCase):
             False, "Error: Is not a dictionary: %s" % (self.data2))
         self.results3 = (False, "Error Message")
 
-    @mock.patch("mysql_db_admin.mongo_libs.ins_doc",
-                mock.Mock(return_value=(False, "Error Message")))
-    def test_mongo_error(self):
-
-        """Function:  test_mongo_error
-
-        Description:  Test with mongo option with an error status.
-
-        Arguments:
-
-        """
-
-        self.assertEqual(
-            mysql_db_admin.data_out(
-                self.data, suppress=self.suppress, mongo=self.mongo,
-                db_tbl=self.db_tbl), self.results3)
-
-    @mock.patch("mysql_db_admin.mongo_libs.ins_doc",
-                mock.Mock(return_value=(True, None)))
-    def test_mongo(self):
-
-        """Function:  test_mongo
-
-        Description:  Test with mongo option.
-
-        Arguments:
-
-        """
-
-        self.assertEqual(
-            mysql_db_admin.data_out(
-                self.data, suppress=self.suppress, mongo=self.mongo,
-                db_tbl=self.db_tbl), self.results)
-
-    @mock.patch("mysql_db_admin.pprint.pprint", mock.Mock(return_value=True))
-    @mock.patch("mysql_db_admin.open", mock.Mock(return_value=True))
+    @mock.patch("mysql_db_admin.gen_libs.write_file",
+                mock.Mock(return_value=True))
     def test_outfile_mode2(self):
 
         """Function:  test_outfile_mode2
@@ -215,8 +178,8 @@ class UnitTest(unittest.TestCase):
                 self.data, suppress=self.suppress, outfile=self.outfile,
                 mode=self.mode2), self.results)
 
-    @mock.patch("mysql_db_admin.pprint.pprint", mock.Mock(return_value=True))
-    @mock.patch("mysql_db_admin.open", mock.Mock(return_value=True))
+    @mock.patch("mysql_db_admin.gen_libs.write_file",
+                mock.Mock(return_value=True))
     def test_outfile_mode(self):
 
         """Function:  test_outfile_mode
@@ -232,8 +195,8 @@ class UnitTest(unittest.TestCase):
                 self.data, suppress=self.suppress, outfile=self.outfile,
                 mode=self.mode), self.results)
 
-    @mock.patch("mysql_db_admin.pprint.pprint", mock.Mock(return_value=True))
-    @mock.patch("mysql_db_admin.open", mock.Mock(return_value=True))
+    @mock.patch("mysql_db_admin.gen_libs.write_file",
+                mock.Mock(return_value=True))
     def test_outfile(self):
 
         """Function:  test_outfile